# compiler/natives/src/strconv: Use js string conversion utilities to improve performance