--- a/compiler/natives/src/strconv/atoi.go
+++ b/compiler/natives/src/strconv/atoi.go
@@ -15,6 +15,19 @@ func Atoi(s string) (int, error) {
 	if len(s) == 0 {
 		return 0, syntaxError(fnAtoi, s)
 	}
+	// Investigate the bytes of the string
+	// Validate each byte is allowed in parsing
+	// Number allows some prefixes that Go does not: "0x" "0b", "0o"
+	// additionally Number accepts decimals where Go does not "10.2"
+	for i := 0; i < len(s); i++ {
+		v := s[i]
+
+		if v < '0' || v > '9' {
+			if v != '+' && v != '-' {
+				return 0, syntaxError(fnAtoi, s)
+			}
+		}
+	}
 	jsValue := js.Global.Call("Number", s, 10)
 	if !js.Global.Call("isFinite", jsValue).Bool() {
 		return 0, syntaxError(fnAtoi, s)
@@ -22,9 +35,9 @@ func Atoi(s string) (int, error) {
 	// Bounds checking
 	floatval := jsValue.Float()
 	if floatval > maxInt32 {
-		return 1<<31 - 1, rangeError(fnAtoi, s)
+		return int(maxInt32), rangeError(fnAtoi, s)
 	} else if floatval < minInt32 {
-		return -1 << 31, rangeError(fnAtoi, s)
+		return int(minInt32), rangeError(fnAtoi, s)
 	}
 	// Success!
 	return jsValue.Int(), nil