Added native implemntations for Atoi and Itoa. This generates smaller and faster code.

Before (using go stdlib)
```
$ gopherjs test strconv --run=Atoi --bench=Atoi
goos: linux
goarch: js
BenchmarkAtoi/Pos/7bit    	 2000000	       606 ns/op
BenchmarkAtoi/Pos/26bit   	 2000000	       628 ns/op
BenchmarkAtoi/Pos/31bit   	  200000	      8700 ns/op
BenchmarkAtoi/Neg/7bit    	 2000000	       660 ns/op
BenchmarkAtoi/Neg/26bit   	 2000000	       735 ns/op
BenchmarkAtoi/Neg/31bit   	  200000	      9030 ns/op
PASS
ok  	strconv	12.110s
```

After (using new js native)
```
$ gopherjs test strconv --run=Atoi --bench=Atoi
goos: linux
goarch: js
BenchmarkAtoi/Pos/7bit    	10000000	       111 ns/op
BenchmarkAtoi/Pos/26bit   	10000000	       150 ns/op
BenchmarkAtoi/Pos/31bit   	10000000	       214 ns/op
BenchmarkAtoi/Neg/7bit    	10000000	       138 ns/op
BenchmarkAtoi/Neg/26bit   	10000000	       158 ns/op
BenchmarkAtoi/Neg/31bit   	10000000	       222 ns/op
PASS
ok  	strconv	11.481s
```

I also wrote a small Itoa benchmark since there wasn't one in std lib. This uses 123456789 as the test number
```
gopherjs test --bench=.
goos: linux
goarch: js
BenchmarkGoItoa      1000000          1217 ns/op
BenchmarkJSItoa     10000000           142 ns/op
PASS
```