@lologarithm 2018-04-05
@shurcooL another optimization PR for you :)